feat: more flexible override for Java REST client (#213)
This PR makes a minor change to allow more flexible override of the Java
REST client `initialize` function. We can use this to override
`initialize` and create an `ApiClient` that uses a custom HTTP client
with mTLS configuration set, for example. Because the member variables
are set with `initApiInstances`, all of the existing methods can be used
with the customized client.

Co-authored-by: Bryan Keller <bkeller@netflix.com>